Resolve cast_lossless pedantic clippy lint
    warning: casts from `u8` to `i32` can be expressed infallibly using `From`
       --> src/lib.rs:367:9
        |
    367 | /         (unsafe {
    368 | |             *EXP_SHIFTS
    369 | |                 .data
    370 | |                 .as_ptr()
    371 | |                 .add((bin_exp + ExpShiftTable::OFFSET) as usize)
    372 | |         }) as i32
        | |_________________^
        |
        = help: an `as` cast can become silently lossy if the types change in the future
        = help: for further information visit https://rust-lang.github.io/rust-clippy/master/index.html#cast_lossless
        = note: `-W clippy::cast-lossless` implied by `-W clippy::pedantic`
        = help: to override `-W clippy::pedantic` add `#[allow(clippy::cast_lossless)]`
    help: use `i32::from` instead
        |
    367 ~         i32::from((unsafe {
    368 +             *EXP_SHIFTS
    369 +                 .data
    370 +                 .as_ptr()
    371 +                 .add((bin_exp + ExpShiftTable::OFFSET) as usize)
    372 +         }))
        |

+++ src/lib.rs
@@ -364,12 +364,12 @@ where
 {
     let num_bits = mem::size_of::<UInt>() * 8;
     if num_bits == 64 && ExpShiftTable::ENABLE && regular {
-        (unsafe {
+        i32::from(unsafe {
             *EXP_SHIFTS
                 .data
                 .as_ptr()
                 .add((bin_exp + ExpShiftTable::OFFSET) as usize)
-        }) as i32
+        })
     } else {
         do_compute_exp_shift(bin_exp, dec_exp)
     }
